refactored boltzmann wealth model to new layout
