+++ examples/boltzmann_wealth_model/run.py
@@ -1,0 +1,3 @@
+from wealth_model.server import server
+
+server.launch()
