+++ examples/boltzmann_wealth_model/wealth_model/__init__.py
@@ -1,0 +1,1 @@
+from .money_model import MoneyModel
